handle differences in types on Linux vs macOS

+++ terminal/cfgetospeed-cfsetospeed/main.c
@@ -24,7 +24,7 @@ int main(void)
         return 1;
     }
 
-    printf("Current output baud rate: %lu\n", output_baud);
+    printf("Current output baud rate: %lu\n", (unsigned long)output_baud);
 
     // Set a new output baud rate
     speed_t new_output_baud = B9600; // 9600 baud
@@ -35,7 +35,7 @@ int main(void)
         return 1;
     }
 
-    printf("New output baud rate: %lu\n", new_output_baud);
+    printf("New output baud rate: %lu\n", (unsigned long)new_output_baud);
 
     // Apply the updated terminal attributes
     if(tcsetattr(STDIN_FILENO, TCSANOW, &term) == -1)
